Added AWS Documentation + Open Charge Map Updater
added swagger documentation for api endpoints, not using request models as they're not available
on serverless offline yet (pr is lodged with the project owners). our current validator uses jsonschema spec so should
be compatible moving forward.

added last status update table collection to remove the need for a dynamo db table scan. Will need to be checked and updated
on insert / update but this is significantly cheaper than scanning the entire collection.